+++ R/mixedAn.default.R
@@ -6,7 +6,9 @@
   
   Ubar <- OL.star <- evi.star <- NULL
   
-  value <- value %||% rep(1, he$n_comparators)/he$n_comparators
+  if (is.null(value)) {
+    value <- value %||% rep(1, he$n_comparators)/he$n_comparators
+  }
   
   Ubar <- compute_Ubar(he, value)
   OL.star <- he$Ustar - Ubar