+++ R/plot.mixedAn.R
@@ -21,10 +21,7 @@
 #' indicating to use the default position and `TRUE` to place it on the
 #' bottom of the plot. Default value is `c(0,1)`, that is in the topleft
 #' corner inside the plot area.
-#' @param graph A string used to select the graphical engine to use for
-#' plotting. Should (partial-)match the two options `"base"` or
-#' `"ggplot2"`. Default value is `"base"`. The `"plotly"` option is not 
-#' implemented for this particular graph.
+#' @template args-graph
 #' @param ...  Arguments to be passed to methods, such as graphical parameters
 #' (see [par()]).
 #' @return \item{evi}{ A ggplot object containing the plot. Returned only if
@@ -33,13 +30,14 @@
 #' cost-effective intervention is included in the market) and the mixed
 #' strategy one (when more than one intervention is considered in the market).
 #' @author Gianluca Baio, Andrea Berardi
-#' @usage plot.mixedAn(x, y.limits=NULL, pos=c(0,1), graph=c("base","ggplot2"),...)
+#' @usage plot.mixedAn(x, y.limits=NULL, pos=c(0,1), graph=options("bcea.graph"),...)
 #' 
 #' @export
 #' 
 plot.mixedAn <- function(x, y.limits = NULL, pos = c(0,1),
-                         graph = c("base","ggplot2"), ...) {
+                         graph = options("bcea.graph"), ...) {
   .Deprecated(new = "evi.plot")
+  graph=unlist(graph) 
   evi.plot(`mixedAn<-`(x),y.limits=NULL, pos=c(0,1),graph=graph,...)
 }